curent calculation
+ calculation of motor current

+++ CAN_motor_controller/Core/Inc/command_interpreter.h
@@ -31,7 +31,7 @@ void transmit_target_motor_speed();
 uint8_t get_temperature_internal();
 uint8_t get_temperature_external();
 uint16_t get_voltage();
-uint16_t get_current();
+uint8_t get_current();
 
 void set_speed(CANbus_msg_t msg);
 void set_max_curr(CANbus_msg_t msg);
